STYLE: solve all the W errors

+++ src/map/free.c
@@ -11,7 +11,7 @@ void free_obs(raccoonmove_t *move)
 {
     int i = 0;
 
-    while (i != move->obs.index_obs) {
+    while (i != (int)move->obs.index_obs) {
         free(move->obs.obstacle[i]);
         free(move->obs.type[i]);
         i++;
